+++ srcs/readline/yank_text_1.c
@@ -14,7 +14,7 @@
 
 int	yank_begin_word(t_matrix *matrix)
 {
-	int col;
+	size_t col;
 
 	col = get_cursor_pos_begin_word(matrix);
 	matrix_string_yank(matrix, matrix->cursor->row, col);
@@ -23,7 +23,7 @@ int	yank_begin_word(t_matrix *matrix)
 
 int	yank_next_word(t_matrix *matrix)
 {
-	int col;
+	size_t col;
 
 	col = get_cursor_pos_next_word(matrix);
 	matrix_string_yank(matrix, matrix->cursor->row, col);
@@ -32,7 +32,7 @@ int	yank_next_word(t_matrix *matrix)
 
 int	yank_end_word(t_matrix *matrix)
 {
-	int col;
+	size_t col;
 
 	col = get_cursor_pos_end_word(matrix);
 	matrix_string_yank(matrix, matrix->cursor->row, col);
@@ -41,7 +41,7 @@ int	yank_end_word(t_matrix *matrix)
 
 int	yank_end(t_matrix *matrix)
 {
-	int col;
+	size_t col;
 
 	col = get_cursor_pos_end(matrix);
 	matrix_string_yank(matrix, matrix->cursor->row, col);
@@ -50,7 +50,7 @@ int	yank_end(t_matrix *matrix)
 
 int	yank_home(t_matrix *matrix)
 {
-	int col;
+	size_t col;
 
 	col = get_cursor_pos_home(matrix);
 	matrix_string_yank(matrix, matrix->cursor->row, col);
